+++ lib/Documents/src/Exceptions/DocumentWithoutFileException.php
@@ -8,17 +8,11 @@ use RZ\Roadiz\Documents\Models\DocumentInterface;
 
 final class DocumentWithoutFileException extends \RuntimeException
 {
-    private DocumentInterface $document;
-
-    public function __construct(DocumentInterface $document)
+    public function __construct(private readonly DocumentInterface $document)
     {
-        $this->document = $document;
         parent::__construct(sprintf('Document (%s) does not have a file on system.', (string) $document));
     }
 
-    /**
-     * @return DocumentInterface
-     */
     public function getDocument(): DocumentInterface
     {
         return $this->document;