+++ lib/DtsGenerator/src/DeclarationGeneratorFactory.php
@@ -17,14 +17,8 @@ use Symfony\Component\HttpFoundation\ParameterBag;
 
 final class DeclarationGeneratorFactory
 {
-    private ParameterBag $nodeTypesBag;
-
-    /**
-     * @param ParameterBag $nodeTypesBag
-     */
-    public function __construct(ParameterBag $nodeTypesBag)
+    public function __construct(private readonly ParameterBag $nodeTypesBag)
     {
-        $this->nodeTypesBag = $nodeTypesBag;
     }
 
     /**
@@ -60,18 +54,12 @@ final class DeclarationGeneratorFactory
      */
     public function createForNodeTypeField(NodeTypeFieldInterface $field): AbstractFieldGenerator
     {
-        switch (true) {
-            case $field->isDocuments():
-                return new DocumentsFieldGenerator($field, $this->nodeTypesBag);
-            case $field->isNodes():
-                return new NodeReferencesFieldGenerator($field, $this->nodeTypesBag);
-            case $field->isChildrenNodes():
-                return new ChildrenNodeFieldGenerator($field, $this->nodeTypesBag);
-            case $field->isMultiple():
-            case $field->isEnum():
-                return new EnumFieldGenerator($field, $this->nodeTypesBag);
-            default:
-                return new ScalarFieldGenerator($field, $this->nodeTypesBag);
-        }
+        return match (true) {
+            $field->isDocuments() => new DocumentsFieldGenerator($field, $this->nodeTypesBag),
+            $field->isNodes() => new NodeReferencesFieldGenerator($field, $this->nodeTypesBag),
+            $field->isChildrenNodes() => new ChildrenNodeFieldGenerator($field, $this->nodeTypesBag),
+            $field->isMultiple(), $field->isEnum() => new EnumFieldGenerator($field, $this->nodeTypesBag),
+            default => new ScalarFieldGenerator($field, $this->nodeTypesBag),
+        };
     }
 }