+++ lib/OpenId/src/Exception/OpenIdConfigurationException.php
@@ -8,13 +8,11 @@ use Throwable;
 
 final class OpenIdConfigurationException extends \RuntimeException
 {
-    /**
-     * @param string $message
-     * @param int $code
-     * @param Throwable|null $previous
-     */
-    public function __construct($message = 'OpenID configuration is not valid', $code = 0, Throwable $previous = null)
-    {
+    public function __construct(
+        string $message = 'OpenID configuration is not valid',
+        int $code = 0,
+        Throwable $previous = null
+    ) {
         parent::__construct($message, $code, $previous);
     }
 }