+++ lib/RoadizRozierBundle/src/Controller/Document/DocumentUnusedController.php
@@ -16,14 +16,8 @@ use Twig\Error\RuntimeError;
 
 final class DocumentUnusedController extends RozierApp
 {
-    private ManagerRegistry $managerRegistry;
-
-    /**
-     * @param ManagerRegistry $managerRegistry
-     */
-    public function __construct(ManagerRegistry $managerRegistry)
+    public function __construct(private readonly ManagerRegistry $managerRegistry)
     {
-        $this->managerRegistry = $managerRegistry;
     }
 
     /**