+++ lib/RoadizRozierBundle/src/Controller/Login/LoginRequestController.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace RZ\Roadiz\RozierBundle\Controller\Login;
 
+use Doctrine\ORM\OptimisticLockException;
+use Doctrine\ORM\ORMException;
 use Psr\Log\LoggerInterface;
 use RZ\Roadiz\CoreBundle\Form\LoginRequestForm;
 use RZ\Roadiz\CoreBundle\Security\User\UserViewer;
@@ -12,20 +14,17 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Themes\Rozier\RozierApp;
+use Twig\Error\RuntimeError;
 
 final class LoginRequestController extends RozierApp
 {
     use LoginRequestTrait;
 
-    private LoggerInterface $logger;
-    private UrlGeneratorInterface $urlGenerator;
-    private UserViewer $userViewer;
-
-    public function __construct(LoggerInterface $logger, UrlGeneratorInterface $urlGenerator, UserViewer $userViewer)
-    {
-        $this->logger = $logger;
-        $this->urlGenerator = $urlGenerator;
-        $this->userViewer = $userViewer;
+    public function __construct(
+        private readonly LoggerInterface $logger,
+        private readonly UrlGeneratorInterface $urlGenerator,
+        private readonly UserViewer $userViewer
+    ) {
     }
 
     protected function getUserViewer(): UserViewer
@@ -37,8 +36,9 @@ final class LoginRequestController extends RozierApp
      * @param Request $request
      *
      * @return Response
-     * @throws \Doctrine\ORM\ORMException
-     * @throws \Doctrine\ORM\OptimisticLockException
+     * @throws RuntimeError
+     * @throws ORMException
+     * @throws OptimisticLockException
      */
     public function indexAction(Request $request): Response
     {
@@ -70,6 +70,7 @@ final class LoginRequestController extends RozierApp
 
     /**
      * @return Response
+     * @throws RuntimeError
      */
     public function confirmAction(): Response
     {