+++ lib/Rozier/src/Controllers/CacheController.php
@@ -47,7 +47,7 @@ final class CacheController extends RozierApp
              */
             return $this->redirectToRoute('adminHomePage');
         }
-
+        $this->prepareBaseAssignation();
         $this->assignation['form'] = $form->createView();
 
         return $this->render('@RoadizRozier/cache/deleteDoctrine.html.twig', $this->assignation);
@@ -77,6 +77,7 @@ final class CacheController extends RozierApp
             return $this->redirectToRoute('adminHomePage');
         }
 
+        $this->prepareBaseAssignation();
         $this->assignation['form'] = $form->createView();
 
         return $this->render('@RoadizRozier/cache/deleteAssets.html.twig', $this->assignation);