+++ lib/Rozier/src/Controllers/CustomForms/CustomFormsUtilsController.php
@@ -4,24 +4,30 @@ declare(strict_types=1);
 
 namespace Themes\Rozier\Controllers\CustomForms;
 
+use Doctrine\Persistence\ManagerRegistry;
 use PhpOffice\PhpSpreadsheet\Exception;
-use RZ\Roadiz\CoreBundle\Entity\CustomForm;
-use RZ\Roadiz\CoreBundle\Entity\CustomFormAnswer;
 use RZ\Roadiz\CoreBundle\CustomForm\CustomFormAnswerSerializer;
-use RZ\Roadiz\CoreBundle\Xlsx\XlsxExporter;
+use RZ\Roadiz\CoreBundle\Entity\CustomForm;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\ResponseHeaderBag;
+use Symfony\Component\HttpFoundation\StreamedResponse;
+use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Contracts\Translation\TranslatorInterface;
 use Themes\Rozier\RozierApp;
 
 class CustomFormsUtilsController extends RozierApp
 {
-    public function __construct(private readonly CustomFormAnswerSerializer $customFormAnswerSerializer)
-    {
+    public function __construct(
+        private readonly ManagerRegistry $managerRegistry,
+        private readonly TranslatorInterface $translator,
+        private readonly CustomFormAnswerSerializer $customFormAnswerSerializer,
+        private readonly SerializerInterface $serializer
+    ) {
     }
 
     /**
-     * Export all custom form's answer in a Xlsx file (.rzt).
+     * Export all custom form's answers in a CSV file.
      *
      * @param Request $request
      * @param int $id
@@ -32,45 +38,35 @@ class CustomFormsUtilsController extends RozierApp
      */
     public function exportAction(Request $request, int $id): Response
     {
-        /** @var CustomForm|null $customForm */
-        $customForm = $this->em()->find(CustomForm::class, $id);
+        $customForm = $this->managerRegistry->getRepository(CustomForm::class)->find($id);
         if (null === $customForm) {
             throw $this->createNotFoundException();
         }
 
         $answers = $customForm->getCustomFormAnswers();
-
-        /**
-         * @var int $key
-         * @var CustomFormAnswer $answer
-         */
+        $answersArray = [];
         foreach ($answers as $key => $answer) {
-            $array = array_merge(
-                [$answer->getIp(), $answer->getSubmittedAt()],
-                $this->customFormAnswerSerializer->toSimpleArray($answer)
-            );
-            $answers[$key] = $array;
+            $answersArray[$key] = $this->customFormAnswerSerializer->toSimpleArray($answer);
         }
 
-        $keys = ["ip", "submitted.date"];
-
         $fields = $customForm->getFieldsLabels();
-        $keys = array_merge($keys, $fields);
+        $keys = [
+            'ip',
+            'submitted.date',
+            ...$fields
+        ];
 
-        $exporter = new XlsxExporter($this->getTranslator());
-        $xlsx = $exporter->exportXlsx($answers, $keys);
-
-        $response = new Response(
-            $xlsx,
-            Response::HTTP_OK,
-            []
-        );
-
+        $response = new StreamedResponse(function () use ($answersArray, $keys) {
+            echo $this->serializer->serialize($answersArray, 'csv', [
+                'csv_headers' => $keys
+            ]);
+        });
+        $response->headers->set('Content-Type', 'text/csv');
         $response->headers->set(
             'Content-Disposition',
             $response->headers->makeDisposition(
                 ResponseHeaderBag::DISPOSITION_ATTACHMENT,
-                $customForm->getName() . '.xlsx'
+                $customForm->getName() . '.csv'
             )
         );
 
@@ -84,15 +80,12 @@ class CustomFormsUtilsController extends RozierApp
      *
      * @param Request $request
      * @param int $id
-     *
      * @return Response
      */
     public function duplicateAction(Request $request, int $id): Response
     {
         $this->denyAccessUnlessGranted('ROLE_ACCESS_CUSTOMFORMS');
-        /** @var CustomForm|null $existingCustomForm */
-        $existingCustomForm = $this->em()->find(CustomForm::class, $id);
-
+        $existingCustomForm = $this->managerRegistry->getRepository(CustomForm::class)->find($id);
         if (null === $existingCustomForm) {
             throw $this->createNotFoundException();
         }
@@ -101,7 +94,7 @@ class CustomFormsUtilsController extends RozierApp
             $newCustomForm = clone $existingCustomForm;
             $newCustomForm->setCreatedAt(new \DateTime());
             $newCustomForm->setUpdatedAt(new \DateTime());
-            $em = $this->em();
+            $em = $this->managerRegistry->getManager();
 
             foreach ($newCustomForm->getFields() as $field) {
                 $em->persist($field);
@@ -110,7 +103,7 @@ class CustomFormsUtilsController extends RozierApp
             $em->persist($newCustomForm);
             $em->flush();
 
-            $msg = $this->getTranslator()->trans("duplicated.custom.form.%name%", [
+            $msg = $this->translator->trans("duplicated.custom.form.%name%", [
                 '%name%' => $existingCustomForm->getDisplayName(),
             ]);
 
@@ -123,7 +116,7 @@ class CustomFormsUtilsController extends RozierApp
         } catch (\Exception $e) {
             $this->publishErrorMessage(
                 $request,
-                $this->getTranslator()->trans("impossible.duplicate.custom.form.%name%", [
+                $this->translator->trans("impossible.duplicate.custom.form.%name%", [
                     '%name%' => $existingCustomForm->getDisplayName(),
                 ]),
                 $newCustomForm