+++ lib/Rozier/src/Explorer/FolderExplorerItem.php
@@ -10,13 +10,10 @@ use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 final class FolderExplorerItem extends AbstractExplorerItem
 {
-    private Folder $folder;
-    private UrlGeneratorInterface $urlGenerator;
-
-    public function __construct(Folder $folder, UrlGeneratorInterface $urlGenerator)
-    {
-        $this->folder = $folder;
-        $this->urlGenerator = $urlGenerator;
+    public function __construct(
+        private readonly Folder $folder,
+        private readonly UrlGeneratorInterface $urlGenerator
+    ) {
     }
 
     /**