+++ lib/Rozier/src/Explorer/SettingExplorerItem.php
@@ -10,13 +10,10 @@ use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 final class SettingExplorerItem extends AbstractExplorerItem
 {
-    private Setting $setting;
-    private UrlGeneratorInterface $urlGenerator;
-
-    public function __construct(Setting $setting, UrlGeneratorInterface $urlGenerator)
-    {
-        $this->setting = $setting;
-        $this->urlGenerator = $urlGenerator;
+    public function __construct(
+        private readonly Setting $setting,
+        private readonly UrlGeneratorInterface $urlGenerator
+    ) {
     }
 
     /**