+++ lib/Rozier/src/Forms/FolderCollectionType.php
@@ -17,14 +17,8 @@ use Themes\Rozier\Explorer\FoldersProvider;
 
 final class FolderCollectionType extends AbstractType
 {
-    protected ManagerRegistry $managerRegistry;
-
-    /**
-     * @param ManagerRegistry $managerRegistry
-     */
-    public function __construct(ManagerRegistry $managerRegistry)
+    public function __construct(private readonly ManagerRegistry $managerRegistry)
     {
-        $this->managerRegistry = $managerRegistry;
     }
 
     /**