+++ lib/Rozier/src/Models/NodeTypeModel.php
@@ -8,14 +8,8 @@ use RZ\Roadiz\CoreBundle\Entity\NodeType;
 
 final class NodeTypeModel implements ModelInterface
 {
-    private NodeType $nodeType;
-
-    /**
-     * @param NodeType $nodeType
-     */
-    public function __construct(NodeType $nodeType)
+    public function __construct(private readonly NodeType $nodeType)
     {
-        $this->nodeType = $nodeType;
     }
 
     public function toArray(): array