+++ lib/Rozier/src/Models/TagModel.php
@@ -9,13 +9,10 @@ use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 final class TagModel implements ModelInterface
 {
-    private Tag $tag;
-    private UrlGeneratorInterface $urlGenerator;
-
-    public function __construct(Tag $tag, UrlGeneratorInterface $urlGenerator)
-    {
-        $this->tag = $tag;
-        $this->urlGenerator = $urlGenerator;
+    public function __construct(
+        private readonly Tag $tag,
+        private readonly UrlGeneratorInterface $urlGenerator
+    ) {
     }
 
     public function toArray(): array